wip: tests reuse
