+++ src/lib/kotlin/slatekit-tests/src/test/kotlin/test/db/Db_Common_Tests.kt
@@ -1,10 +1,59 @@
 package test.db
 
 import kiit.common.data.IDb
+import org.junit.Assert
 import org.junit.Test
 
 abstract class Db_Common_Tests : DbTestCases {
     abstract fun db(): IDb
+
+    @Test
+    override fun can_insert_sql_raw() {
+        val db = db()
+        val id = db.insertGetId(insertSqlRaw()).toLong()
+        Assert.assertTrue(id > 0L)
+    }
+
+
+    @Test
+    override fun can_update() {
+        val db = db()
+        // 1. add
+        val id = db.insert(insertSqlRaw())
+        Assert.assertTrue(id > 0)
+
+        // 2. update
+        val sqlUpdate = "update ${table()} set test_int = 987 where id = $id"
+        val count = db.update(sqlUpdate)
+        Assert.assertTrue(count > 0)
+
+        // 3. get
+        val sql = "select test_int from ${table()} where id = $id"
+        val updatedVal = db.getScalarInt(sql, null)
+        Assert.assertTrue(updatedVal == 987)
+    }
+
+    @Test
+    override fun can_delete() {
+        val db = db()
+        // 1. add
+        val id = db.insert(insertSqlRaw())
+        Assert.assertTrue(id > 0)
+
+        // 2. get count
+        val sqlCount = "select count(*) from ${table()} where id = $id;"
+        val countBefore = db.getScalarInt(sqlCount)
+        Assert.assertTrue(countBefore == 1)
+
+        // 3. delete
+        val sql = "delete from ${table()} where id = $id;"
+        db.execute(sql)
+
+        // 4. get count after delete
+        val countAfter = db.getScalarInt(sqlCount)
+        Assert.assertTrue(countAfter == 0)
+    }
+
 
     @Test
     override fun can_query_scalar_string() {
@@ -74,4 +123,7 @@ abstract class Db_Common_Tests : DbTestCases {
             Db_Fixtures.localDateTime
         )
     }
+
+    abstract fun table(): String
+    abstract fun insertSqlRaw(): String
 }
