+++ src/lib/kotlin/slatekit-tests/src/test/kotlin/test/db/Db_Postgres_Tests.kt
@@ -29,10 +29,9 @@ import java.util.*
  * 1. Docker: at root of folder : {root}/docker-compose.yml
  * 2. MySql : See credentials in the docker file
  */
-class Db_Postgres_Tests : DbTestCases {
+class Db_Postgres_Tests : Db_Common_Tests(), DbTestCases {
 
-    private fun db(): IDb = EntitySetup.db(Vendor.Postgres)
-
+    override fun db(): IDb = EntitySetup.db(Vendor.Postgres)
 
 
     @Test
@@ -45,7 +44,7 @@ class Db_Postgres_Tests : DbTestCases {
     @Test
     override fun can_insert_sql_raw() {
         val db = db()
-        val id = db.insertGetId(insertSqlRaw).toLong()
+        val id = db.insertGetId(insertSqlRaw()).toLong()
         Assert.assertTrue(id > 0L)
     }
 
@@ -88,16 +87,16 @@ class Db_Postgres_Tests : DbTestCases {
         val db = db()
 
         // 1. add
-        val id = db.insert(insertSqlRaw)
+        val id = db.insert(insertSqlRaw())
         Assert.assertTrue(id > 0)
 
         // 2. update
-        val sqlUpdate = """update $table set test_int = 987 where "id" = $id;"""
+        val sqlUpdate = """update ${table()} set test_int = 987 where "id" = $id;"""
         val count = db.update(sqlUpdate)
         Assert.assertTrue(count > 0)
 
         // 3. get
-        val sql = """select test_int from $table where "id" = $id;"""
+        val sql = """select test_int from ${table()} where "id" = $id;"""
         val updatedVal = db.getScalarInt(sql, null)
         Assert.assertTrue(updatedVal == 987)
     }
@@ -106,11 +105,11 @@ class Db_Postgres_Tests : DbTestCases {
     override fun can_get() {
         val db = db()
         // 1. add
-        val id = db.insert(insertSqlRaw)
+        val id = db.insert(insertSqlRaw())
         Assert.assertTrue(id > 0)
 
         // 2. update
-        val sqlGet = """select * from $table where "id" = $id;"""
+        val sqlGet = """select * from ${table()} where "id" = $id;"""
         val item = db.mapOne(sqlGet, null) { rec ->
             val longid = rec.getLong("id")
             SampleEntityImmutable(
@@ -140,91 +139,21 @@ class Db_Postgres_Tests : DbTestCases {
     override fun can_delete() {
         val db = db()
         // 1. add
-        val id = db.insert(insertSqlRaw)
+        val id = db.insert(insertSqlRaw())
         Assert.assertTrue(id > 0)
 
         // 2. get count
-        val sqlCount = "select count(*) from $table where id = $id;"
+        val sqlCount = "select count(*) from ${table()} where id = $id;"
         val countBefore = db.getScalarInt(sqlCount)
         Assert.assertTrue(countBefore == 1)
 
         // 3. delete
-        val sql = "delete from $table where id = $id;"
+        val sql = "delete from ${table()} where id = $id;"
         db.execute(sql)
 
         // 4. get count after delete
         val countAfter = db.getScalarInt(sqlCount)
         Assert.assertTrue(countAfter == 0)
-    }
-
-
-    @Test
-    override fun can_query_scalar_string() {
-        ensure_scalar("test_string", { db, sql -> db.getScalarString(sql, null) }, "abc")
-    }
-
-
-    @Test
-    override fun can_query_scalar_bool() {
-        ensure_scalar("test_bool", { db, sql -> db.getScalarBool(sql, null) }, true)
-    }
-
-
-    @Test
-    override fun can_query_scalar_short() {
-        ensure_scalar("test_short", { db, sql -> db.getScalarShort(sql, null) }, 123)
-    }
-
-
-    @Test
-    override fun can_query_scalar_int() {
-        ensure_scalar("test_int", { db, sql -> db.getScalarInt(sql, null) }, 123456)
-    }
-
-
-    @Test
-    override fun can_query_scalar_long() {
-        ensure_scalar("test_long", { db, sql -> db.getScalarLong(sql, null) }, 123456789)
-    }
-
-
-    @Test
-    override fun can_query_scalar_float() {
-        ensure_scalar("test_float", { db, sql -> db.getScalarFloat(sql, null) }, 123.45f)
-    }
-
-
-    @Test
-    override fun can_query_scalar_double() {
-        ensure_scalar("test_double", { db, sql -> db.getScalarDouble(sql, null) }, 123456.789)
-    }
-
-
-    @Test
-    override fun can_query_scalar_localdate() {
-        ensure_scalar("test_localdate", { db, sql -> db.getScalarLocalDate(sql, null) }, Db_Fixtures.localDate)
-    }
-
-
-    @Test
-    override fun can_query_scalar_localtime() {
-        ensure_scalar("test_localtime", { db, sql -> db.getScalarLocalTime(sql, null) }, Db_Fixtures.localTime)
-    }
-
-
-    @Test
-    override fun can_query_scalar_localdatetime() {
-        ensure_scalar(
-            "test_localdatetime",
-            { db, sql -> db.getScalarLocalDateTime(sql, null) },
-            Db_Fixtures.localDateTime
-        )
-    }
-
-
-    @Test
-    override fun can_query_scalar_date() {
-        ensure_scalar("test_localdatetime", { db, sql -> db.getScalarZonedDateTime(sql, null) }, Db_Fixtures.zonedDateTime)
     }
 
 
@@ -246,8 +175,8 @@ class Db_Postgres_Tests : DbTestCases {
 
     override fun <T> ensure_scalar(colName: String, callback: (IDb, String) -> T, expected: T): Unit {
         val db = db()
-        val id = db.insert(insertSqlRaw)
-        val sql = "select $colName from $table where id = $id;"
+        val id = db.insert(insertSqlRaw())
+        val sql = "select $colName from ${table()} where id = $id;"
         val actual = callback(db, sql)
         println(actual)
         println(expected)
@@ -255,7 +184,7 @@ class Db_Postgres_Tests : DbTestCases {
     }
 
 
-    private val table get() = """"unit_tests"."sample_entity""""
+    override fun table():String = """"unit_tests"."sample_entity""""
 
     private val insertSqlPrep = """
             INSERT INTO "unit_tests"."sample_entity"
@@ -275,7 +204,7 @@ class Db_Postgres_Tests : DbTestCases {
             );
         """
 
-    private val insertSqlRaw = """
+    override fun insertSqlRaw():String = """
             INSERT INTO "unit_tests"."sample_entity"
             (
                 "test_string"      , "test_string_enc"  , "test_bool",
